requirements.txt
streamlit
streamlit-authenticator
google-cloud-storage
watchdog
pandas
plotly
jupyterlab
numpy
scikit-learn